requirements.txt
arxiv==2.1.0
requests==2.31.0
pyyaml==6.0.1
volcengine-python-sdk[ark]==1.0.98
supabase==2.3.0
resend==2.0.0
python-dateutil==2.9.0
tenacity==8.2.3
loguru==0.7.2
